Press F1 in save search to open help text + document new advanced search
I tried to document everything the advanced search does, but it was mostly a guessing game discovering all its features. It appears to be insanely powerful, supporting AND, OR, negation, parenthesis,quote-encapsulated phrases, and limiting search to only certain fields. These can be chained together as well. I tried giving examples that could possibly be useful.

Terms like before: and after: are not part of advanced search, and thus are global search options even if they're inside parenthesis