Fix unnatural stacking near powers of 2 positions
Displacing the swapped-out particle by the difference between the integer positions (derived from the float position via rounding) of the particles would sometimes cause the floats in Particle to switch to a lower precision range, and this would overflow them such that the integer position would change by an amount different from what was intended (exactly the difference between the integer positions of the particles). Such switches between precision ranges can only happen when a sufficiently big positive number is added to a positive float near the positive end of its current precision range. Swapping displacements are usually small, so this problem mostly manifested near the positive end of precision ranges, that is, near powers of 2.

Informally, for illustrative purposes, consider the float position (either x or y, doesn't matter) 511.49995, fairly close to the positive end of the precision range [256, 512), and the swap displacement +1: these yield the position 512.49995. Assume that this is not be representable, and that the two nearest representable values are 512.4998 and 512.5. 512.49995 is closer to 512.5, so that's what it gets rounded to. But oops, 511 = round(511.49995) + 1 != round(512.5) = 513, so we're now overwriting the wrong pmap entry.

Reproduce with

	sim.paused(true)
	sim.clearSim()
	local brmt = sim.partCreate(-1, 128, 100, elem.DEFAULT_PT_BRMT)
	local dust = sim.partCreate(-1, 127, 101, elem.DEFAULT_PT_DUST)
	sim.createBox(125, 98, 130, 103, elem.DEFAULT_PT_DMND)
	sim.partProperty(dust, "x", 127.49999)
	sim.frameRender(2)

The BRMT and the DUST should swap places; instead, without the fix, the DUST gets stacked under a DMND.

This commit fixes this problem by moving the swapped-out particle exactly where the moving particle is.