Fix crash on exit while there are Lua components visible
Broken since 02b679aec37f and extremely similar to 74386631e0b6, which makes sure that both LSI and GameView are alive when destroying Lua windows: in this case, they both need to be alive when destroying components attached to the main window. Also, the main window has to exist still.