Simplify event handler management
Semantics are mostly the same:

 - event.register pushes a handler at the end of the handler chain, such that it has the lowest priority
 - event.unregister removes at most one handler, which is the first one that matches
 - a handler failing causes it to be removed from the chain

But removing handlers from the chain doesn't mess up handler order for the frame anymore: until this commit, removing a handler from the chain would shift all handlers in the chain while it was being traversed, which meant that some handlers were always skipped whenever any were removed.

Validate this with the following script:

	local t = {}
	for i = 1, 10 do
		table.insert(t, i)
	end
	for i = #t - 1, 1, -1 do
		local j = math.random(1, i)
		t[i], t[j] = t[j], t[i]
	end
	print(table.concat(t, " "))
	event.register(event.MOUSEDOWN, function()
		print("=====")
	end)
	for i = 1, #t do
		local c = t[i]
		local function f()
			print(i)
			c = c - 1
			if c == 0 then
				event.unregister(event.MOUSEDOWN, f)
			end
		end
		event.register(event.MOUSEDOWN, f)
	end

The expected behaviour is that numbered event handlers are removed once they've been run as many times as the correspondingly numbered entry in t dictates. This, among other things, means that exactly one event handler is removed every MOUSEDOWN. This was absolutely not the case until now; removals were all over the place.