Use working directory as the Path for freedesktop shortcuts
Our freedesktop shortcuts had been using the path to the executable instead ever since they were added in 2df333c99f2e. This doesn't make sense because the working directory can be different from the path to the executable, and results in the hard-to-debug situation where TPT potentially picks up a different user profile if started with a shortcut from the one it picks up when run directly.