Disable ALPN for socket.tcp
libcurl defaults to setting up ALPN for various versions of HTTP, but since socket.tcp isn't used for this, this should be disabled.

TODO: Enable configuring a custom list of ALPN protocols. The problem with this is that libcurl doesn't support this on its own, and asking the TLS backend it uses directly is problematic because it can come with one of many, and supporting all of them is a pain. All official builds use mbedtls, but supporting ALPN configuration only on official builds and on local builds that link libcurl that in turn links mbedtls is not ideal.